+++ cryptography/src/bls12381/benches/kzg_commit.rs
@@ -1,6 +1,8 @@
 use commonware_cryptography::{
-    bls12381::primitives::group::{Scalar, G1},
-    kzg::{open, setup::Ethereum},
+    bls12381::{
+        kzg::{commit, setup::Ethereum},
+        primitives::group::{Scalar, G1},
+    },
 };
 use criterion::{criterion_group, BatchSize, Criterion};
 use rand::rngs::OsRng as rng;
@@ -8,9 +10,10 @@ use std::hint::black_box;
 
 const DEGREES: &[usize] = &[64, 256, 1024, 4096];
 
-fn benchmark_open(c: &mut Criterion) {
+fn benchmark_commit(c: &mut Criterion) {
     let setup = Ethereum::new();
 
+    // Benchmark for different polynomial degrees
     for &degree in DEGREES {
         c.bench_function(&format!("{}/degree={degree}", module_path!()), |b| {
             b.iter_batched(
@@ -19,11 +22,10 @@ fn benchmark_open(c: &mut Criterion) {
                     for _ in 0..degree {
                         coeffs.push(Scalar::from_rand(&mut rng));
                     }
-                    let point = Scalar::from_rand(&mut rng);
-                    (coeffs, point)
+                    coeffs
                 },
-                |(coeffs, point)| {
-                    black_box(open::<Ethereum, G1>(&coeffs, &point, &setup).unwrap());
+                |coeffs| {
+                    black_box(commit::<Ethereum, G1>(&coeffs, &setup).unwrap());
                 },
                 BatchSize::SmallInput,
             );
@@ -31,4 +33,4 @@ fn benchmark_open(c: &mut Criterion) {
     }
 }
 
-criterion_group!(benches, benchmark_open);
+criterion_group!(benches, benchmark_commit);
