--- cryptography/src/kzg/benches/bench.rs
@@ -1,7 +1,0 @@
-use criterion::criterion_main;
-
-mod commit;
-mod open;
-mod verify;
-
-criterion_main!(commit::benches, open::benches, verify::benches);
