+++ cryptography/src/lib.rs
@@ -32,7 +32,6 @@ pub use bloomfilter::BloomFilter;
 pub mod handshake;
 pub mod lthash;
 pub use lthash::LtHash;
-pub mod kzg;
 pub mod secp256r1;
 pub mod transcript;
